+++ sdk/neonpostgres/arm-neonpostgres/samples/v2-beta/javascript/organizationsListBySubscriptionSample.js
@@ -1,0 +1,29 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { PostgresClient } = require("@azure/arm-neonpostgres");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to list OrganizationResource resources by subscription ID
+ *
+ * @summary list OrganizationResource resources by subscription ID
+ * x-ms-original-file: 2025-06-23-preview/Organizations_ListBySubscription_MaximumSet_Gen.json
+ */
+async function organizationsListBySubscriptionMaximumSet() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "DFF26289-4E9C-46D0-890E-F8BE27BDA8C2";
+  const client = new PostgresClient(credential, subscriptionId);
+  const resArray = new Array();
+  for await (const item of client.organizations.listBySubscription()) {
+    resArray.push(item);
+  }
+
+  console.log(resArray);
+}
+
+async function main() {
+  await organizationsListBySubscriptionMaximumSet();
+}
+
+main().catch(console.error);