+++ sdk/neonpostgres/arm-neonpostgres/src/api/computes/index.ts
@@ -1,11 +1,5 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export { list, $delete, update, createOrUpdate, get } from "./operations.js";
-export {
-  ComputesListOptionalParams,
-  ComputesDeleteOptionalParams,
-  ComputesUpdateOptionalParams,
-  ComputesCreateOrUpdateOptionalParams,
-  ComputesGetOptionalParams,
-} from "./options.js";
+export { list } from "./operations.js";
+export { ComputesListOptionalParams } from "./options.js";