+++ sdk/neonpostgres/arm-neonpostgres/src/api/neonRoles/index.ts
@@ -1,11 +1,9 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export { list, $delete, update, createOrUpdate, get } from "./operations.js";
+export { $delete, createOrUpdate, list } from "./operations.js";
 export {
-  NeonRolesListOptionalParams,
   NeonRolesDeleteOptionalParams,
-  NeonRolesUpdateOptionalParams,
   NeonRolesCreateOrUpdateOptionalParams,
-  NeonRolesGetOptionalParams,
+  NeonRolesListOptionalParams,
 } from "./options.js";